auth_samples_python/requirements.txt
Flask==3.0.3
Authlib==1.3.1
python-dotenv==1.0.1
